# ARROW-13170  [C++] Reducing branching in compute/kernels/vector_selection.cc