This PR adds the changes discussed in ARROW-13170.