@nirandaperera 2021-07-07
These were the perf results from my local desktop. 
https://gist.github.com/nirandaperera/dfafb77865e948514ca520162be10558