@wesm 2021-07-08
I guess branch prediction works really well on modern processors — when the filter values are mostly false then the if block is rarely executed. In the 50% selected case, branch prediction doesn't help so this method yields speedups. 

Not sure if it's worth investigating, but moving `out_position_` from a class member to a stack variable could have some performance impact. Another thought is removing some of the offset arithmetic

Lastly, since we're using `SetBitTo` a lot, it might be worth checking the superscalar variant described at

https://graphics.stanford.edu/~seander/bithacks.html#ConditionalSetOrClearBitsWithoutBranching