@nirandaperera 2021-07-08
@wesm with super scalar variant, I get the following,
```
BEFORE:
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Non-regressions: (10)
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
                benchmark         baseline        contender  change %                                    counters
     FilterInt64FilterNoNulls/1048576/4    1.423 GiB/sec    2.658 GiB/sec    86.837  {'run_name': 'FilterInt64FilterNoNulls/1048576/4', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 1021, 'data null%': 0.1, 'mask null%': 0.0, 'select%': 50.0}
     FilterInt64FilterNoNulls/1048576/7    1.367 GiB/sec    2.398 GiB/sec    75.409   {'run_name': 'FilterInt64FilterNoNulls/1048576/7', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 978, 'data null%': 1.0, 'mask null%': 0.0, 'select%': 50.0}
    FilterInt64FilterNoNulls/1048576/13    1.318 GiB/sec    2.223 GiB/sec    68.611 {'run_name': 'FilterInt64FilterNoNulls/1048576/13', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 936, 'data null%': 90.0, 'mask null%': 0.0, 'select%': 50.0}
    FilterInt64FilterNoNulls/1048576/10    1.323 GiB/sec    2.220 GiB/sec    67.824 {'run_name': 'FilterInt64FilterNoNulls/1048576/10', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 950, 'data null%': 10.0, 'mask null%': 0.0, 'select%': 50.0}
    FilterInt64FilterNoNulls/1048576/12    6.890 GiB/sec    7.092 GiB/sec     2.938{'run_name': 'FilterInt64FilterNoNulls/1048576/12', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 4949, 'data null%': 90.0, 'mask null%': 0.0, 'select%': 99.9}
     FilterInt64FilterNoNulls/1048576/1    2.358 GiB/sec    2.386 GiB/sec     1.166  {'run_name': 'FilterInt64FilterNoNulls/1048576/1', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 1676, 'data null%': 0.0, 'mask null%': 0.0, 'select%': 50.0}
     FilterInt64FilterNoNulls/1048576/6    8.110 GiB/sec    8.034 GiB/sec    -0.940  {'run_name': 'FilterInt64FilterNoNulls/1048576/6', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 5779, 'data null%': 1.0, 'mask null%': 0.0, 'select%': 99.9}
     FilterInt64FilterNoNulls/1048576/2   39.687 GiB/sec   39.170 GiB/sec    -1.301  {'run_name': 'FilterInt64FilterNoNulls/1048576/2', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 28002, 'data null%': 0.0, 'mask null%': 0.0, 'select%': 1.0}
     FilterInt64FilterNoNulls/1048576/0   14.103 GiB/sec   13.782 GiB/sec    -2.278  {'run_name': 'FilterInt64FilterNoNulls/1048576/0', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 9845, 'data null%': 0.0, 'mask null%': 0.0, 'select%': 99.9}
     FilterInt64FilterNoNulls/1048576/9    7.006 GiB/sec    6.800 GiB/sec    -2.951 {'run_name': 'FilterInt64FilterNoNulls/1048576/9', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 5061, 'data null%': 10.0, 'mask null%': 0.0, 'select%': 99.9}

---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Regressions: (5)
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
                             benchmark        baseline       contender  change %                                     counters
    FilterInt64FilterNoNulls/1048576/3   9.850 GiB/sec   9.107 GiB/sec    -7.538   {'run_name': 'FilterInt64FilterNoNulls/1048576/3', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 7039, 'data null%': 0.1, 'mask null%': 0.0, 'select%': 99.9}
    FilterInt64FilterNoNulls/1048576/5  10.072 GiB/sec   4.863 GiB/sec   -51.715    {'run_name': 'FilterInt64FilterNoNulls/1048576/5', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 7212, 'data null%': 0.1, 'mask null%': 0.0, 'select%': 1.0}
    FilterInt64FilterNoNulls/1048576/8   9.578 GiB/sec   4.409 GiB/sec   -53.966    {'run_name': 'FilterInt64FilterNoNulls/1048576/8', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 6873, 'data null%': 1.0, 'mask null%': 0.0, 'select%': 1.0}
   FilterInt64FilterNoNulls/1048576/11   9.509 GiB/sec   4.080 GiB/sec   -57.096  {'run_name': 'FilterInt64FilterNoNulls/1048576/11', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 6842, 'data null%': 10.0, 'mask null%': 0.0, 'select%': 1.0}
   FilterInt64FilterNoNulls/1048576/14   9.532 GiB/sec   4.084 GiB/sec   -57.154  {'run_name': 'FilterInt64FilterNoNulls/1048576/14', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 6748, 'data null%': 90.0, 'mask null%': 0.0, 'select%': 1.0}


AFTER:
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Non-regressions: (9)
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
                          benchmark       baseline      contender  change %                                     counters
 FilterInt64FilterNoNulls/1048576/4  1.422 GiB/sec  2.976 GiB/sec   109.349   {'run_name': 'FilterInt64FilterNoNulls/1048576/4', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 1017, 'data null%': 0.1, 'mask null%': 0.0, 'select%': 50.0}
 FilterInt64FilterNoNulls/1048576/7  1.363 GiB/sec  2.500 GiB/sec    83.386    {'run_name': 'FilterInt64FilterNoNulls/1048576/7', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 978, 'data null%': 1.0, 'mask null%': 0.0, 'select%': 50.0}
FilterInt64FilterNoNulls/1048576/13  1.320 GiB/sec  2.216 GiB/sec    67.878  {'run_name': 'FilterInt64FilterNoNulls/1048576/13', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 945, 'data null%': 90.0, 'mask null%': 0.0, 'select%': 50.0}
FilterInt64FilterNoNulls/1048576/10  1.322 GiB/sec  2.201 GiB/sec    66.533  {'run_name': 'FilterInt64FilterNoNulls/1048576/10', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 947, 'data null%': 10.0, 'mask null%': 0.0, 'select%': 50.0}
 FilterInt64FilterNoNulls/1048576/1  2.351 GiB/sec  2.627 GiB/sec    11.739   {'run_name': 'FilterInt64FilterNoNulls/1048576/1', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 1683, 'data null%': 0.0, 'mask null%': 0.0, 'select%': 50.0}
 FilterInt64FilterNoNulls/1048576/0 13.413 GiB/sec 13.631 GiB/sec     1.628   {'run_name': 'FilterInt64FilterNoNulls/1048576/0', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 9605, 'data null%': 0.0, 'mask null%': 0.0, 'select%': 99.9}
 FilterInt64FilterNoNulls/1048576/6  7.672 GiB/sec  7.593 GiB/sec    -1.027   {'run_name': 'FilterInt64FilterNoNulls/1048576/6', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 5715, 'data null%': 1.0, 'mask null%': 0.0, 'select%': 99.9}
 FilterInt64FilterNoNulls/1048576/2 39.856 GiB/sec 38.819 GiB/sec    -2.600   {'run_name': 'FilterInt64FilterNoNulls/1048576/2', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 28575, 'data null%': 0.0, 'mask null%': 0.0, 'select%': 1.0}
FilterInt64FilterNoNulls/1048576/12  6.806 GiB/sec  6.558 GiB/sec    -3.653 {'run_name': 'FilterInt64FilterNoNulls/1048576/12', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 4785, 'data null%': 90.0, 'mask null%': 0.0, 'select%': 99.9}

----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Regressions: (6)
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
                          benchmark       baseline     contender  change %                                    counters
 FilterInt64FilterNoNulls/1048576/9  6.843 GiB/sec 6.472 GiB/sec    -5.426 {'run_name': 'FilterInt64FilterNoNulls/1048576/9', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 4689, 'data null%': 10.0, 'mask null%': 0.0, 'select%': 99.9}
 FilterInt64FilterNoNulls/1048576/3  9.378 GiB/sec 7.717 GiB/sec   -17.710  {'run_name': 'FilterInt64FilterNoNulls/1048576/3', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 6703, 'data null%': 0.1, 'mask null%': 0.0, 'select%': 99.9}
 FilterInt64FilterNoNulls/1048576/5 10.064 GiB/sec 5.390 GiB/sec   -46.442   {'run_name': 'FilterInt64FilterNoNulls/1048576/5', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 7216, 'data null%': 0.1, 'mask null%': 0.0, 'select%': 1.0}
 FilterInt64FilterNoNulls/1048576/8  9.600 GiB/sec 4.489 GiB/sec   -53.233   {'run_name': 'FilterInt64FilterNoNulls/1048576/8', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 6865, 'data null%': 1.0, 'mask null%': 0.0, 'select%': 1.0}
FilterInt64FilterNoNulls/1048576/11  9.529 GiB/sec 4.105 GiB/sec   -56.926 {'run_name': 'FilterInt64FilterNoNulls/1048576/11', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 6825, 'data null%': 10.0, 'mask null%': 0.0, 'select%': 1.0}
FilterInt64FilterNoNulls/1048576/14  9.537 GiB/sec 4.098 GiB/sec   -57.027 {'run_name': 'FilterInt64FilterNoNulls/1048576/14', 'repetitions': 0, 'repetition_index': 0, 'threads': 1, 'iterations': 6825, 'data null%': 90.0, 'mask null%': 0.0, 'select%': 1.0}
```